% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_function.R
\name{clusterFunc}
\alias{clusterFunc}
\title{rgrumps cluster function}
\usage{
clusterFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\value{
grumpsObj
}
\description{
This function clusters a distance matrix and returns a hierarchical clustering object
}
\examples{
clusterFunc(grumpsObj)
}
\keyword{grumps}
