% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorDendrogram.R
\name{ColorDendrogram}
\alias{ColorDendrogram}
\title{rgrumps dendrogram function}
\usage{
ColorDendrogram(
  hc,
  y,
  main = "",
  branchlength = 0.7,
  labels = NULL,
  xlab = NULL,
  sub = NULL,
  ylab = "",
  cex.main = NULL
)
}
\arguments{
\item{hc}{output of hclust}

\item{y}{vector containing labels of hclust output}

\item{main}{title of plot}

\item{branchlength}{how long colored region is of branch}

\item{labels}{labels for tips}

\item{xlab}{x-axis label}

\item{sub}{sub-x-axis label}

\item{ylab}{y-axis label}

\item{cex.main}{amount to enlarge main title of figure}
}
\description{
This function creates a color dendrogram from hc.
}
\examples{
ColorDendrogram(grumpsObj$hClust, grumpsObj$mcl)
}
\keyword{grumps}
