% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_function.R
\name{dataframeFunc}
\alias{dataframeFunc}
\title{rgrumps dataframe function}
\usage{
dataframeFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\value{
grumpsObj
}
\description{
This function reads a .csv file output by GRUMPS and saves it as a matrix
}
\examples{
dataframeFunc(grumpsObj)
}
\keyword{grumps}
