% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_function.R
\name{dendrogramFunc}
\alias{dendrogramFunc}
\title{rgrumps dendrogram output function}
\usage{
dendrogramFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\description{
This function creates a dendrogram from a grumpsObj and saves it to a file
}
\examples{
dendrogramFunc(grumpsObj)
}
\keyword{grumps}
