% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grumps_function.R
\name{grumpsFunc}
\alias{grumpsFunc}
\title{rgrumps grumps function}
\usage{
grumpsFunc(
  filepath,
  mode = "heatmap",
  cutoff = 0.0125,
  clusteringmethod = "ward.D2",
  tree = "yes"
)
}
\arguments{
\item{filepath}{Filepath to a .csv file output from GRUMPS}

\item{mode}{Mode to run rgrumps in. Options are: 'heatmap', 'dendrogram', or 'general'. Default: 'heatmap'.}

\item{cutoff}{Cutoff to use for creating subgroups from hclust output. If mode is not 'general', the max height of the hclust output is multiplied by this value. Default: 1.25E-02.}

\item{clusteringmethod}{Clustering method to use by hclust. Available methods are: 'ward.D', 'ward.D2', 'single', 'complete', 'average', 'mcquitty', 'median', or 'centroid'. Default: 'ward.D2'.}

\item{tree}{Output the resultant dendrogram as a .nwk file. Default: 'yes'.}
}
\value{
A grumpsObj, which is used by all other functions in this package.
}
\description{
This function creates a grumpsObj
}
\examples{
grumpsFunc('/filepath/to/distmat.csv')
grumpsFunc('/filepath/to/distmat.csv', mode='heatmap')
grumpsFunc('/filepath/to/distmat.csv', mode='heatmap', cutoff=1.25E-02,clusteringmethod='ward.D2',tree='yes')
}
\keyword{grumps}
