% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_function.R
\name{heatmapFunc}
\alias{heatmapFunc}
\title{rgrumps heatmap function}
\usage{
heatmapFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\description{
This function creates a heatmap from a grumpsObj and saves it to a file
}
\examples{
heatmapFunc(grumpsObj)
}
\keyword{grumps}
