% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightcutter_function.R
\name{heightCutter}
\alias{heightCutter}
\title{rgrumps tree height cutter function}
\usage{
heightCutter(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\value{
grumpsObj
}
\description{
This function cuts a tree from a grumpsObj using the user supplied cutoff
}
\examples{
heightCutter(grumpsObj)
}
\keyword{grumps}
