% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelout_function.R
\name{labeloutFunc}
\alias{labeloutFunc}
\title{rgrumps heatmap function}
\usage{
labeloutFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\description{
This function outputs the labels and groups from a grumpsObj
}
\examples{
labeloutFunc(grumpsObj)
}
\keyword{grumps}
