% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl_function.R
\name{mclFunc}
\alias{mclFunc}
\title{rgrumps tree cutting function}
\usage{
mclFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumps_function}
}
\value{
grumpsObj
}
\description{
This function cuts the grumps tree and returns a vector of group membership
}
\examples{
mclFunc(grumpsObj)
}
\keyword{grumps}
