% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_function.R
\name{treeFunc}
\alias{treeFunc}
\title{rgrumps tree function}
\usage{
treeFunc(grumpsObj)
}
\arguments{
\item{grumpsObj}{grumpsObj from grumpsFunction}
}
\description{
This function outputs the dendrogram as a .nwk file
}
\examples{
treeFunc(grumpsObj)
}
\keyword{grumps}
